RFA: moved most classes into their own files...
