+++ exercise.main/Discount/DiscountManager.cs
@@ -1,16 +1,16 @@
-﻿namespace exercise.main
+﻿namespace exercise.main.Discount
 {
-    public class DiscountManager 
+    public class DiscountManager
     {
-        
+
         List<Discount> discountTypes = new List<Discount>();
         //public void addDiscountType<T>(params object[] args) where T : Discount
-        private Inventory _inventory; 
+        private Inventory _inventory;
         public DiscountManager(Inventory inventory)
         {
-            this._inventory = inventory;
+            _inventory = inventory;
         }
-        public void addDiscountType(Discount discount) 
+        public void addDiscountType(Discount discount)
         {
             // TODO: Check for and remove identicals...
             discountTypes.Add(discount);
@@ -21,20 +21,20 @@
             // Looks for conflicting deals, removes them and favors best value deals
 
             List<Tuple<int, int, string>> conflictingIndexes = new List<Tuple<int, int, string>>();
-            List<DiscountedProductCount> possibleCombinations= new List<DiscountedProductCount>();
+            List<DiscountedProductCount> possibleCombinations = new List<DiscountedProductCount>();
 
             // Count/collect possible conflicting deals
             for (int i = 0; i < possibleDiscounts.Count; i++)
             {
                 bool noConflict = true;
-                for (int j = i+1; j < possibleDiscounts.Count; j++)
+                for (int j = i + 1; j < possibleDiscounts.Count; j++)
                 {
                     foreach (var key in possibleDiscounts[j].SKU_amount.Keys)
                     {
 
-                        if(possibleDiscounts[i].SKU_amount.ContainsKey(key))
+                        if (possibleDiscounts[i].SKU_amount.ContainsKey(key))
                         {
-                            conflictingIndexes.Add(new (i, j, key));
+                            conflictingIndexes.Add(new(i, j, key));
                             noConflict = false;
                         }
                     }
@@ -45,7 +45,7 @@
                 }
             }
 
-            foreach(var c in conflictingIndexes)
+            foreach (var c in conflictingIndexes)
             {
                 int index_i = c.Item1;
                 int index_j = c.Item2;
@@ -63,7 +63,7 @@
                     possibleCombinations.Add(possibleDiscounts[index_j]);
 
                 }
-                
+
 
                 //defprice.
 
@@ -93,16 +93,16 @@
             // Remove Discounted products from the product list...
             foreach (var discount in bestDealsDiscounts)
             {
-                 
-                foreach ( var di in discount.SKU_amount)
+
+                foreach (var di in discount.SKU_amount)
                 {
-                    var sku = di.Key; 
+                    var sku = di.Key;
                     int amount = di.Value;
 
                     int counted = 0;
                     List<BaseProduct> temp = new List<BaseProduct>();
-                    
-                    foreach ( var product in productList)
+
+                    foreach (var product in productList)
                     {
                         if (sku != product.SKU || counted >= amount)
                         {
@@ -113,30 +113,30 @@
                             counted++;
                         }
                     }
-                    
-                    productList = temp; 
 
-                    
+                    productList = temp;
+
+
                 }
             }
 
             //Dictionary<string, Tuple<int, float>> nameAmountPrice = new Dictionary<string, Tuple<int, float>>();
             Dictionary<string, OrderData> nameAmountPrice = new Dictionary<string, OrderData>();
-            
+
             foreach (var product in bestDealsDiscounts)
             {
                 //string nameStr = "deal: " + string.Join(" + ", product.SKU_amount.Keys);
-                string nameStr = "deal: " + string.Join(" + ", product.SKU_amount.Keys.Select(x => this._inventory.getName(x)));
-                
+                string nameStr = "deal: " + string.Join(" + ", product.SKU_amount.Keys.Select(x => _inventory.getName(x)));
+
                 nameAmountPrice[nameStr] = new OrderData
                 {
                     name = nameStr,
-                    amount = product.discountMultiple, 
+                    amount = product.discountMultiple,
                     individual_price = 0.0f, // TODO: fix ...
                     discounted_price = product.finalPrice,
                     total_price = product.finalPrice,
                     saving = product.possibleSavings,
-                    
+
                 };
             }
 
@@ -157,9 +157,9 @@
                 //float? defPrice = basket.getDefaultPrice(product.Key);
                 //float? defPrice = this._inventory.getPrice(product.Key);
                 //if (defPrice == null)
-                    //Debug.Assert(false, "defPrice cant be zero...");
+                //Debug.Assert(false, "defPrice cant be zero...");
 
-                float defPrice = this._inventory.getPrice(product.Key);
+                float defPrice = _inventory.getPrice(product.Key);
 
                 //nameAmountPrice[product.Key] = new (product.Value, defPrice.Value);
                 nameAmountPrice[product.Key] = new OrderData
@@ -172,7 +172,7 @@
                     //total_price = defPrice.Value * product.Value,
                     total_price = defPrice * product.Value,
                     saving = 0.0f,
-                }; 
+                };
             }
 
             return nameAmountPrice;
