+++ exercise.main/Discount/DiscountedProductCount.cs
@@ -1,0 +1,14 @@
+﻿namespace exercise.main.Discount
+{
+    public class DiscountedProductCount
+    {
+        public Dictionary<string, int> SKU_amount = new Dictionary<string, int>();
+        public int discountMultiple = 0;
+        public float possibleSavings = 0.0f;
+        public float finalPrice = 0.0f;
+    }
+
+
+
+
+}
