+++ exercise.main/Order.cs
@@ -1,0 +1,15 @@
+﻿namespace exercise.main
+{
+    public class Order
+    {
+        public Dictionary<string, OrderData> orderDatas;
+        public Order(Basket basket)
+        {
+
+        }
+    }
+
+
+
+
+}
