+++ exercise.main/OrderData.cs
@@ -1,0 +1,16 @@
+﻿namespace exercise.main
+{
+    public struct OrderData
+    {
+        public string name;
+        public int amount;
+        public float individual_price; 
+        public float discounted_price; 
+        public float total_price; 
+        public float saving; 
+    }
+
+
+
+
+}
